+++ components/style/Trending.module.css
