+++ components/test/Footer.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Footer from './Footer';
+import Footer from '../main/Footer';
 
 describe('# Footer Component', () => {
   it('renders without crashing', () => {
