+++ components/test/Search.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, act } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Search from './Search';
+import Search from '../main/search/Search';
 
 describe('Search component', () => {
   it('Renders Search compnent', () => {
