+++ components/test/Trending.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, act } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Trending from './Search';
+import Trending from '../main/nav/Trending';
 
 describe('Trending component', () => {
   it('Renders Trending compnent', () => {
